feat(Popover): Add `Popover.Trigger`
