+++ packages/react/src/components/Popover/Popover.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryFn } from '@storybook/react';
-import React, { useRef, useState } from 'react';
+import React from 'react';
 
 import { Button, Paragraph } from '../..';
 
@@ -17,27 +17,11 @@ export default {
 } as Meta;
 
 export const Preview: StoryFn<typeof Popover> = (args) => {
-  const buttonRef = useRef<HTMLButtonElement | null>(null);
-  const [open, setOpen] = useState(false);
-
   return (
-    <>
-      <Button
-        ref={buttonRef}
-        onClick={() => setOpen(!open)}
-        aria-expanded={open}
-      >
-        My trigger
-      </Button>
-      <Popover
-        {...args}
-        open={open || args.open}
-        onClose={() => setOpen(false)}
-        anchorEl={buttonRef.current}
-      >
-        <Popover.Content>popover content</Popover.Content>
-      </Popover>
-    </>
+    <Popover {...args}>
+      <Popover.Trigger>My trigger!</Popover.Trigger>
+      <Popover.Content>popover content</Popover.Content>
+    </Popover>
   );
 };
 
@@ -51,50 +35,45 @@ Preview.args = {
 Preview.decorators = [marginDecorator];
 
 export const Variants: StoryFn<typeof Popover> = () => {
-  const buttonRef = useRef<HTMLButtonElement | null>(null);
-  const [open, setOpen] = useState(false);
-
-  React.useEffect(() => {
-    setOpen(true);
-  }, []);
-
   return (
     <>
-      <Button
-        ref={buttonRef}
-        aria-expanded={open}
-      >
-        My trigger
-      </Button>
       <Popover
-        anchorEl={buttonRef.current}
-        open={open}
+        open={true}
         placement='top'
       >
+        <Popover.Trigger asChild>
+          <span>popover</span>
+        </Popover.Trigger>
         <Popover.Content>default</Popover.Content>
       </Popover>
       <Popover
-        anchorEl={buttonRef.current}
-        open={open}
-        placement='right'
+        open={true}
+        placement='bottom'
         variant='danger'
       >
+        <Popover.Trigger asChild>
+          <span>popover</span>
+        </Popover.Trigger>
         <Popover.Content>danger</Popover.Content>
       </Popover>
       <Popover
-        anchorEl={buttonRef.current}
-        open={open}
-        placement='bottom'
+        open={true}
+        placement='top'
         variant='info'
       >
+        <Popover.Trigger asChild>
+          <span>popover</span>
+        </Popover.Trigger>
         <Popover.Content>info</Popover.Content>
       </Popover>
       <Popover
-        anchorEl={buttonRef.current}
-        open={open}
-        placement='left'
+        open={true}
+        placement='bottom'
         variant='warning'
       >
+        <Popover.Trigger asChild>
+          <span>popover</span>
+        </Popover.Trigger>
         <Popover.Content>warning</Popover.Content>
       </Popover>
     </>
@@ -104,30 +83,15 @@ export const Variants: StoryFn<typeof Popover> = () => {
 Variants.decorators = [marginDecorator];
 
 export const InteractiveContent: StoryFn<typeof Popover> = () => {
-  const buttonRef = useRef<HTMLButtonElement | null>(null);
-  const [open, setOpen] = useState(false);
-
   return (
     <>
-      <Button
-        ref={buttonRef}
-        onClick={() => setOpen(!open)}
-        aria-expanded={open}
-      >
-        My trigger
-      </Button>
-      <Popover
-        anchorEl={buttonRef.current}
-        placement='top'
-        open={open}
-        onClose={() => setOpen(false)}
-      >
+      <Popover>
+        <Popover.Trigger>My trigger</Popover.Trigger>
         <Popover.Content>
           <Paragraph>Er du sikker på at du vil slette?</Paragraph>
           <Button
             size='small'
             color='danger'
-            onClick={() => setOpen(false)}
           >
             Slett
           </Button>
@@ -140,27 +104,10 @@ export const InteractiveContent: StoryFn<typeof Popover> = () => {
 InteractiveContent.decorators = [marginDecorator];
 
 export const InPortal = () => {
-  const buttonRef = useRef<HTMLButtonElement | null>(null);
-  const [open, setOpen] = useState(false);
-
   return (
-    <>
-      <Button
-        ref={buttonRef}
-        onClick={() => setOpen(!open)}
-        aria-expanded={open}
-      >
-        My trigger
-      </Button>
-      <Popover
-        anchorEl={buttonRef.current}
-        placement='top'
-        open={open}
-        onClose={() => setOpen(false)}
-        portal
-      >
-        <Popover.Content>popover content</Popover.Content>
-      </Popover>
-    </>
+    <Popover portal>
+      <Popover.Trigger>My trigger</Popover.Trigger>
+      <Popover.Content>popover content</Popover.Content>
+    </Popover>
   );
 };
